flask_app/src/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
pymongo==4.3.3
Werkzeug==2.2.3
python-dotenv==1.0.0
pandas==1.5.3
requests==2.28.2